Use modal composable
